force box version 1.0.1
Newer versions of this box bundle a version of Puppet which no longer supports
the `manifestdir` parameter that we are using:

==> default: Error: Could not parse application options: invalid option: --manifestdir

As a temporary fix, this patch ensures that the box version is one which still
supports the parameter. This should be sufficient enough until I figure out an
appropriate change which would allow us to upgrade.